moved lstm logics inside stock_analytics app
